Simplify calculating domains for single charts.
